Lift numOfElements state up to BarVisualizer()
numOfElements state is now in BarVisualizer() in BarVisualizer.js as a
shared state and passed as props to the slider and BarVisualizerBody.
This allows the slider component to change the number of elements/bars
and have the body re render to reflect this.

This change currently disables the working buttons in the toolbar.