Change cell view algo buttons behavior
Prevent array from resetting when and algorithm button is pressed during
the step through of a previous algorithm. This essentially triggers a
sorting algorithm on the currently displayed array.

Change allows users to run the algorithms after they have already sorted
it to see behavior. Side effect is that users can run a new sort on the
displayed array during a step through of a previous sort.